+++ api/src/routes/admin.rs
@@ -107,7 +107,7 @@ pub struct UserDetail {
     pub is_active: bool,
     pub is_admin: bool,
     pub webhook_limit: i32,
-    pub webhook_count: i32,
+    pub webhook_count: i64,
     pub created_at: chrono::DateTime<chrono::Utc>,
     pub endpoints: Vec<EndpointSummary>,
     pub recent_deliveries: Vec<DeliverySummary>,
